refactor(app): CHooseRobotToRunProtocolSlideout->Desktop
