+++ app/src/organisms/Desktop/ChooseProtocolSlideout/__tests__/ChooseProtocolSlideout.test.tsx
@@ -16,13 +16,13 @@ import {
   storedProtocolDataWithoutRunTimeParameters,
 } from '/app/redux/protocol-storage/__fixtures__'
 import { useTrackCreateProtocolRunEvent } from '/app/organisms/Desktop/Devices/hooks'
-import { useCreateRunFromProtocol } from '/app/organisms/ChooseRobotToRunProtocolSlideout/useCreateRunFromProtocol'
+import { useCreateRunFromProtocol } from '/app/organisms/Desktop/ChooseRobotToRunProtocolSlideout/useCreateRunFromProtocol'
 import { ChooseProtocolSlideout } from '../'
 import { useNotifyDataReady } from '/app/resources/useNotifyDataReady'
 import type { ProtocolAnalysisOutput } from '@opentrons/shared-data'
 
 vi.mock(
-  '/app/organisms/ChooseRobotToRunProtocolSlideout/useCreateRunFromProtocol'
+  '/app/organisms/Desktop/ChooseRobotToRunProtocolSlideout/useCreateRunFromProtocol'
 )
 vi.mock('/app/redux/protocol-storage')
 vi.mock('/app/organisms/Desktop/Devices/hooks')
