+++ app/src/organisms/Desktop/ChooseProtocolSlideout/index.tsx
@@ -52,7 +52,7 @@ import { ToggleButton } from '/app/atoms/buttons'
 import { MiniCard } from '/app/molecules/MiniCard'
 import { UploadInput } from '/app/molecules/UploadInput'
 import { useTrackCreateProtocolRunEvent } from '/app/organisms/Desktop/Devices/hooks'
-import { useCreateRunFromProtocol } from '/app/organisms/ChooseRobotToRunProtocolSlideout/useCreateRunFromProtocol'
+import { useCreateRunFromProtocol } from '/app/organisms/Desktop/ChooseRobotToRunProtocolSlideout/useCreateRunFromProtocol'
 import { ApplyHistoricOffsets } from '/app/organisms/ApplyHistoricOffsets'
 import { useOffsetCandidatesForAnalysis } from '/app/organisms/ApplyHistoricOffsets/hooks/useOffsetCandidatesForAnalysis'
 import { FileCard } from '/app/organisms/ChooseRobotSlideout/FileCard'
