+++ app/src/organisms/ProtocolDetails/__tests__/ProtocolDetails.test.tsx
@@ -5,7 +5,7 @@ import { describe, it, beforeEach, vi, expect, afterEach } from 'vitest'
 
 import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
-import { ChooseRobotToRunProtocolSlideout } from '/app/organisms/ChooseRobotToRunProtocolSlideout'
+import { ChooseRobotToRunProtocolSlideout } from '/app/organisms/Desktop/ChooseRobotToRunProtocolSlideout'
 import {
   useTrackEvent,
   ANALYTICS_PROTOCOL_PROCEED_TO_RUN,
@@ -33,7 +33,7 @@ vi.mock('/app/redux/analytics')
 vi.mock('/app/redux/custom-labware/selectors')
 vi.mock('/app/redux/discovery/selectors')
 vi.mock('/app/redux/protocol-storage/selectors')
-vi.mock('/app/organisms/ChooseRobotToRunProtocolSlideout')
+vi.mock('/app/organisms/Desktop/ChooseRobotToRunProtocolSlideout')
 vi.mock('/app/organisms/SendProtocolToFlexSlideout')
 
 const render = (
