+++ app/src/organisms/ProtocolDetails/index.tsx
@@ -60,7 +60,7 @@ import {
   analyzeProtocol,
 } from '/app/redux/protocol-storage'
 import { useFeatureFlag } from '/app/redux/config'
-import { ChooseRobotToRunProtocolSlideout } from '../ChooseRobotToRunProtocolSlideout'
+import { ChooseRobotToRunProtocolSlideout } from '/app/organisms/Desktop/ChooseRobotToRunProtocolSlideout'
 import { SendProtocolToFlexSlideout } from '../SendProtocolToFlexSlideout'
 import { ProtocolAnalysisFailure } from '../ProtocolAnalysisFailure'
 import { ProtocolStatusBanner } from '../ProtocolStatusBanner'
