+++ app/src/organisms/ProtocolsLanding/ProtocolList.tsx
@@ -29,7 +29,7 @@ import {
 } from '/app/redux/config'
 import { useSortedProtocols } from './hooks'
 import { Slideout } from '/app/atoms/Slideout'
-import { ChooseRobotToRunProtocolSlideout } from '../ChooseRobotToRunProtocolSlideout'
+import { ChooseRobotToRunProtocolSlideout } from '/app/organisms/Desktop/ChooseRobotToRunProtocolSlideout'
 import { SendProtocolToFlexSlideout } from '../SendProtocolToFlexSlideout'
 import { ProtocolUploadInput } from './ProtocolUploadInput'
 import { ProtocolCard } from './ProtocolCard'
